updated add-remove.js and utils.js files
