+++ webpack.config.js
@@ -17,7 +17,7 @@ module.exports = {
     static: {
       directory: path.resolve(__dirname, 'dist'),
     },
-    port: 4001,
+    port: 4002,
     open: true,
     hot: true,
     compress: true,
